Improve unit tests for file descriptor handling.
